// src/app/model/emprego.ts
export class Emprego {

    status: number;
    id: number;
    localTrabalho: string;
    tipoLocalTrabalho: number;
    ocupacao: string;
    mesInicial: number;
    mesFinal: number;
    anoInicial: number;
    anoFinal: number;
    pais: number;
    flUltimoEmprego: boolean;
    ultimoSalario: number;
    motivoSaida: string;
    ultimoTrabalho: string;
    periodoUltimoTrabalho: string;

}